refactor(frontend): auth store cleanup
