+++ packages/frontend/components/Borrow/CollateralDropdown.tsx
@@ -1,7 +1,7 @@
 import React from "react"
 import { ListItemIcon, ListItemText, MenuItem, Typography } from "@mui/material"
 
-import { chains } from "../../store/auth.store"
+import { chains } from "../../services/chains"
 import NetworkIcon from "../NetworkIcon"
 
 type Chain = (typeof chains)[0]
