+++ packages/frontend/components/Borrow/Overview.tsx
@@ -25,7 +25,7 @@ import { BorrowingVault } from "@x-fuji/sdk"
 import CurrencyCard from "./CurrencyCard"
 import LTVProgressBar from "./LTVProgressBar"
 import ClickableTooltip from "../Layout/ClickableTooltip"
-import { DEFAULT_LTV_RECOMMENDED } from "../../consts/borrow"
+import { DEFAULT_LTV_RECOMMENDED } from "../../constants/borrow"
 import ProviderIcon from "../ProviderIcon"
 import NetworkIcon from "../NetworkIcon"
 import { useBorrow } from "../../store/borrow.store"
