+++ packages/frontend/components/Borrow/RoutingModal.tsx
@@ -5,7 +5,7 @@ import CloseIcon from "@mui/icons-material/Close"
 
 import RouteCard from "./RouteCard"
 import { useBorrow } from "../../store/borrow.store"
-import { chainName } from "../../helpers/chainName"
+import { chainName } from "../../services/chains"
 import NetworkIcon from "../NetworkIcon"
 import TokenIcon from "../TokenIcon"
