+++ packages/frontend/components/Borrow/TransactionProcessingModal.tsx
@@ -28,7 +28,7 @@ import { RoutingStep } from "@x-fuji/sdk"
 import NetworkIcon from "../NetworkIcon"
 import { useHistory } from "../../store/history.store"
 import { formatUnits } from "ethers/lib/utils"
-import { chainName } from "../../helpers/chainName"
+import { chainName } from "../../services/chains"
 import { transactionLink } from "../../helpers/transactionInformations"
 import { AddTokenButton } from "../AddTokenButton"
 import { useAuth } from "../../store/auth.store"
