+++ packages/frontend/components/Markets/MarketsTable.tsx
@@ -13,12 +13,12 @@ import {
 import InfoOutlinedIcon from "@mui/icons-material/InfoOutlined"
 import MarketsTableRow from "./MarketsTableRow"
 import { useEffect, useState } from "react"
-import { sdk } from "../../store/auth.store"
 import {
   BorrowingVaultWithFinancials,
   LendingProviderDetails,
 } from "@x-fuji/sdk"
-import { chainName } from "../../helpers/chainName"
+import { chainName } from "../../services/chains"
+import { sdk } from "../../services/sdk"
 import { SizableTableCell } from "../SizableTableCell"
 
 export type Row = {
