+++ packages/frontend/store/borrow.store.ts
@@ -11,12 +11,14 @@ import {
 } from "@x-fuji/sdk"
 import { formatUnits, parseUnits } from "ethers/lib/utils"
 import produce, { setAutoFreeze } from "immer"
-import create from "zustand"
+import { create } from "zustand"
 import { debounce } from "debounce"
 
-import { sdk, testChains, useAuth } from "./auth.store"
-import { Position } from "./Position"
-import { DEFAULT_LTV_MAX, DEFAULT_LTV_TRESHOLD } from "../consts/borrow"
+import { useAuth } from "./auth.store"
+import { Position } from "./models/Position"
+import { testChains } from "../services/chains"
+import { sdk } from "../services/sdk"
+import { DEFAULT_LTV_MAX, DEFAULT_LTV_TRESHOLD } from "../constants/borrow"
 import { ethers, Signature } from "ethers"
 import { toHistoryRoutingStep, useHistory } from "./history.store"
 import { useSnack } from "./snackbar.store"
