+++ packages/frontend/store/history.store.ts
@@ -6,9 +6,9 @@ import {
   Token,
 } from "@x-fuji/sdk"
 import produce from "immer"
-import create from "zustand"
+import { create } from "zustand"
 import { devtools, persist } from "zustand/middleware"
-import { sdk } from "./auth.store"
+import { sdk } from "../services/sdk"
 import ethers from "ethers"
 import { useSnack } from "./snackbar.store"
 import { formatUnits } from "ethers/lib/utils"
