+++ packages/frontend/store/snackbar.store.ts
@@ -1,5 +1,5 @@
 import produce from "immer"
-import create from "zustand"
+import { create } from "zustand"
 import { devtools } from "zustand/middleware"
 
 export type SnackStore = SnackState & SnackActions
